+++ src/utils.ts
@@ -1,4 +1,12 @@
-export type ElemType<A> = A extends Array<infer Item> ? Item : never
+interface NotAnArrayType<_T> {
+  readonly _: unique symbol
+}
+
+export type ElemType<A> = IsOptional<A> extends true
+  ? NotAnArrayType<A>
+  : A extends (infer Item)[]
+  ? Item
+  : NotAnArrayType<A>
 
 // Conduct the check through return value types to work around the
 // Distributive Conditional Types feature:
@@ -11,3 +19,12 @@ export type Eq<A, B> = (() => A) extends () => B
 
 // A if it's equal to B, otherwise B
 export type Simplify<A, B> = Eq<A, B> extends true ? A : B
+
+export type IsOptional<A> = Or<ExtendsUndefined<A>, ExtendsNull<A>>
+
+type ExtendsUndefined<A> = Eq<A | undefined, A>
+type ExtendsNull<A> = Eq<A | null, A>
+
+type Or<A extends true | false, B extends true | false> = A extends true
+  ? true
+  : B