+++ CHANGELOG.md
@@ -6,6 +6,9 @@
   - Add `Getter`, `AffineFold` and `Fold` optic classes
   - Add `.to()` method for creating Getters
 
+- **Fixes**
+  - Disallow array optics on `T[] | undefined` and `T[] | null`
+
 ## 0.1.0
 
 - Initial release